docs: use Note alert to highlight WIP
This uses GitHub's "alert" syntax for highlighting the fact that the README is
more of a spec.

This is to draw attention to it a bit better (and maybe to experiment with the
syntax a bit + deliberatly try to be more familiar with it). Not sure exactly
which level should be used - but avoiding the stronger levels to already start
from the point of avoiding their overuse