feat!: download OpenTTD to a cache directory
Only trying to support Linux and MacOS, because that's all I have access to at
the moment.

These tests don't mock anything, and really do download OpenTTD. Maybe not
ideal, but it's the easiest thing for now.

BREAKING CHANGE: This drops support for Python earlier than 3.8. This is
because the latest version of platformdirs only support Python 3.8 onwards.
Often I think using an earlier version is fine, but because it's involved in
saving files to the user's machine, I'm opting to just use the most recent out
of paranoia.

This should also mean we can now pin the dev dependencies in CI - they were
unpinned to allow different versions to run earlier versions of Python.