docs: tweak accessibility attributes
The SVGs in the badges themselves have accessibility-related attributes and
tags, so I don't think we need them.

But have added `role="img"` to the SVG of the OpenTTDLab logo, which I think
identifies all the items in the SVG as a single image. Have also corrected the
name of the package inside our logo to be consistent with how it's styled
elsewhere.