docs: put copyright/license notice in SVG logo
This is my best-effort at abiding by the GPL v2 that according to
https://commons.wikimedia.org/wiki/File:Openttdlogo.svg the OpenTTD was
originally licensed under.

Note that https://commons.wikimedia.org/wiki/File:Openttdlogo.svg credits the
logo to "OpenTTD Project", but at https://www.openttd.org/ and
https://github.com/OpenTTD/OpenTTD/blob/master/CREDITS.md it uses the word
"team" rather than "project".

Dates are not included in the list of modifications in the SVG file. Because
this is in source control, then hopefully the dates there are enough to
satisfy a requirement of GPL v2:

> a) You must cause the modified files to carry prominent notices stating that
> you changed the files and the date of any change.