feat: ability to fetch AIs published on content server
There are a few steps involved here. Hopefully it's as simple as it can
reasonably be.

The client code does take both a name and unique ID of the AI. Ideally we would
just take name so it's clear to the user which AI is being used. However, names
are not unique, so to disambiguate them just in case there are multiple AIs
with the same name, it also requires the unique ID