feat: run experiments in parallel
This adds the ability to run experiments in parallel, using the max_workers
argument to run_experiment. The name of this arugument is chosen for
consistency with the concurrent.futures Python standard library module.

It defaults to os.cpu_count(): OpenTTD I think is single threaded, and the
settings I _think_ should take it to ~100% of a CPU. I suspect could do with
some metrics to see what a good default really is.

This does use the ThreadPoolExecutor because for the most part the Python part
doesn't do much - OpenTTD runs in a subprocess started from each of these
threads. Parsing the saved games is maybe an exception, and probably won't be
parallelised because of the GIL. However, I suspect it's a tiny fraction of
overall runtime in most cases, and it's unlikely that each run of OpenTTD
finishes at exactly the same time.