feat: start AIs immediately using a startup script
This starts the AIs in the game immediately using a startup script, rather than
waiting for them to be launched by the game engine.

Although not quite possible here yet, this is a step towards passing parameters
to AIs, which would be useful when using OpenTTDLab to choose appropriate
default parameters for them.

BREAKING CHANGE: Changes the results from experiments

This changes the results of OpenTTDLab because the AIs start at a different
point in time than they did before. But hopefully this is a better choice -
they start immediately.