feat: slightly less clunky API for AIs
This makes it slightly easier to pass AIs to the API - no nested tuples, and no
requirements for an empty parameter list for example.

BREAKING CHANGE: How AIs were passed to the run_experiment function has changed.

It was an iterable of tuples, but now an iterable of the return values of the 4
AI-fetcher functions.