feat: option to take final screenshot
This adds the option to take a screenshot at the end of each simulation, by the
`final_screenshot_directory` option to `run_experiment`.

How it works is that it loads OpenTTD again at the end of each simulation, but
with the normal video driver, loading the final autosave file, takes a
screenshot, and then exits. This can result in the OpenTTD window being visible
for a few moments for each run, This seems unavoidable at the moment - it
doesn't seem possible to take screenshots using the null video driver.