feat!: rename run_experiment to run_experiments
I think it seems more natural that a single "experiment" is essentially a
single run of OpenTTD, and a call to this function runs a lot of them.

This I think this will work especially with a future change that will have
`experiments` as a single parameters that controlls all of these experiments.

BREAKING CHANGE: This renames run_experiment to run_experiments