feat!: more flexible batching of experiments
This replaces several parameters with the combined 'experiments' parameter.
This means a single call to run_experiment allows, for example, a single call
to run_experiment to create creates for different sets of ai_params.

How these experiments can be distinguished in the results remains to be done.

BREAKING CHANGE: Several parameters have changed in the core run_exeriment
function