feat: allow OpenTTD config set per run_experiments
This allows, for example, a single call to run_experiments to cover multiple
terrain types, which are set via config.

It also removes the "base" detail from the name. It was chosen to emphasise
that under the hood, OpenTDDLab adds to it, but I think this detail isn't
really needed to be emphasised - it adds to the load of what you have to think
about somewhat needlessly.