feat: add result processor
From running experiments with many runs, memory usage would get higher and
higher and higher, eventually causing Python to get killed because it would use
loads of memory (e.g. I have seen it use 30GB). Suspecting that keeping all the
parsed save games in memory is a contributing factor - usually only want to
keep a very small amount of data, especially when running for a long time I
know what data I want to extract because I've run smaller "pilot" experiments.

This change adds an intermediate "result processor" that means that, hopefully,
only what's needed is stored.

Also, am suspecting that parsing the savegame which is done in Python is making
the scaling not great due to Python's GIL. If we move to a multi-processing
solution that could address that, but also would be good to have a way of
reducing inter-process transfer, and having such a processor could do that.