feat: make the result processor return iterable
This changes how the result processor works - it now returns an iterable of
rows that will be returned in the final results set.

This means it can return a result row per player for example, which adds a bit
of flexibility and I think makes it possible to return results that are more
(Hadley Wickham) tidy.