feat: output combined stdout and stderr, and indicate if we think it errored
This adds "output" to each result row, which was the combined text stdout and
stderr output for that run of OpenTTD - if there is no error this is usually
empty.

It also adds an "error" key which indicates if a script errored during that
run, and exited, which means it can be excluded from results/investigated
further as appropriate.