refactor: step towards automatically downloading dependencies
To do this the download function is going to have to return some sort of list
of everything that an AI needs, in order to then be copied into the run
directory.

So returning just the one thing the download function is downloading, and then
using it, is a small step towards this.