feat: uncompress binaries into the run directory rather than the cache directory
For some reason in older releases of OpenTTD, when uncompressing the dmg files
to where there were already files, it would result in "Operation not permitted"
errors (and only on the second run or beyond).

To solve this, uncompressing the files into the run directory, which gets
cleared up automatically at the end anyway.

This also means the cache directory is left tidier, with no ucompressed files,
which I don't think is bad thing.