feat: move to multiprocess from threads
Doing some very light nosing around `top` when experiments are running, often
there are fewer openttd processing running than expected, and Python is 100%.
So I suspect that Python is parsing the savegame files, and the GIL stops good
scaling.

So, moving to using processes instead of threads for the parsing component.

Also, using the multiprocess module, with dill, rather than Python's built in
multiprocessing, because it allows more things to be transferred to and from
the process (e.g. lambda and local functions).