refactor: AI and AI library funcs return iterables of data
The AI and AI library funcs now return iterables of data, and specifically the
_bananas_download function does this (rather than writing to a target
file).

This is a step towards a nicer API for it ready to make it public. And,
hopefully, be able to expose more data for reproducible experiments, so the
exact same version from bananas is used again later.