feat: allow download of exact version from BaNaNaS
This change allows client code to fetch the exact version of content from
BaNaNaS, if the full MD5 is supplied.

This change also now returns the full MD5 of content downloaded, but only for
content that has been distributed with a license that I think allows download
of the version in perpetuity. This covers virtually all current content on
BaNaNaS. The only content in BaNaNaS right now where this doesn't is one with a
"Custom" license.