feat!: simplify configuration of where to fetch OpenTTD binaries from
This combines openttd_base_url and opengfx_base_url into a single parameter,
openttd_cdn_url, which now should not include the path/folder component.

This is a stepping stone to being able to easily use nightly versions of
OpenTTD. To specify their location another parameter could have been
introduced, but I'm not sure of the use case. Keeping it simple, opting to just
have the one parameter that specifies the location of the OpenTTD CDN as a
"just in case"/"let's keep it flexible/configurable" parameter.