feat: support some OpenTTD and OpenGFX nightlies
This includes removing the sha256 of each of the binary file archives from the
location of the binary files inside the "run" directory.

This is to avoid the error on Windows:

> FileNotFoundError: [WinError 206] The filename or extension is too long

The "run" directory is already a temporary directory with at least one unique
ID, and we already have checked the sha256 of the file contents by this point,
so having the sha256 of the binary archive in the path doesn't give anything.

Note that although this supports some nightlies from a downloading and running
OpenTTD point of view, the version of OpenTTD must still be compatible with
OpenTTDLab in terms of extracting data, and not all version of OpenTTD are
compatible.